r-and-d/decision-tree-generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-tree-generator"
version = "0.1.0"
description = "Deterministic, reproducible decision tree generation from YAML"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MCP Comparison Project"}
]
keywords = ["decision-tree", "yaml", "mermaid", "graphviz", "html"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
dt-mermaid = "decision_tree.cli:mermaid_main"
dt-graphviz = "decision_tree.cli:graphviz_main"
dt-html = "decision_tree.cli:html_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["decision_tree*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
